% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/guessGender.R
\name{getListElement}
\alias{getListElement}
\title{Get an element from a list}
\usage{
getListElement(listName, elementName)
}
\description{
Helper function that returns NA instead of NULL when a missing list element
is requested, otherwise returns the element itself.
}
\keyword{internal}

