% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/guessGender.R
\name{guessGender}
\alias{guessGender}
\title{Guess names' genders}
\usage{
guessGender(nameVector, countryCode = NA, languageCode = NA, apiKey = NA)
}
\arguments{
\item{nameVector}{A vector containing one or more names to look up.}

\item{countryCode}{An optional ISO 3166-1 alpha-2 country code.}

\item{languageCode}{An optional ISO 639-1 language code. Only one of
countryCode or languageCode can be specified.}

\item{apiKey}{An optional API key for genderize.io.}
}
\description{
This function uses the genderize.io API to supply estimates of the gender one
or more names.
}
\examples{
guessGender(c("Eamon", "Sean"), countryCode = "US")
}

