% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/guessGender.R
\name{lookupNameVectorGenderize}
\alias{lookupNameVectorGenderize}
\title{Look up a vector of names on genderize.io.}
\usage{
lookupNameVectorGenderize(nameVector, countryCode = NA, languageCode = NA,
  apiKey = NA)
}
\arguments{
\item{nameVector}{A vector containing one or more names to look up.}

\item{countryCode}{An optional ISO 3166-1 alpha-2 country code.}

\item{languageCode}{An optional ISO 639-1 language code. Only one of
countryCode or languageCode can be specified.}

\item{apiKey}{An optional API key for genderize.io.}
}
\description{
This function actually implements the genderize.io API. Can only query 10
names at a time.
}
\keyword{internal}

