% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/guessGender.R
\name{checkLanguageCountryCodes}
\alias{checkLanguageCountryCodes}
\title{Check country and language code.}
\usage{
checkLanguageCountryCodes(countryCode, languageCode)
}
\description{
Makes sure that no more than one of countryCode or languageCode is *not* NA
(i.e., they can both be NA, or one can be NA).
}
\keyword{internal}

